fix: return tags along with other artwork data

+++ backend/src/main/java/com/kyeeego/digitalportfolio/application/service/FileStorageServiceImpl.java
@@ -28,8 +28,9 @@ public class FileStorageServiceImpl implements FileStorageService {
 
         try {
             Files.createDirectories(storageLocation);
-            Files.copy(this.getClass().getClassLoader().getResourceAsStream("defaultpfp.jpg"),
-                    storageLocation.resolve("defaultpfp.jpg"));
+            if (Files.notExists(storageLocation.resolve("defaultpfp.jpg")))
+                Files.copy(this.getClass().getClassLoader().getResourceAsStream("defaultpfp.jpg"),
+                        storageLocation.resolve("defaultpfp.jpg"));
         } catch (Exception ex) {
             throw new RuntimeException(
                     "Could not create the directory where the uploaded files will be stored.", ex);
